+++ src/seedsigner/helpers/bcur.py
