+++ src/seedsigner/helpers/camera_process.py
