+++ src/seedsigner/helpers/path.py
