+++ src/seedsigner/helpers/qr.py
