+++ src/seedsigner/helpers/ur2/bytewords.py
