+++ src/seedsigner/helpers/ur2/constants.py
