+++ src/seedsigner/helpers/ur2/ur_encoder.py
