+++ src/seedsigner/models/__init__.py
@@ -1,0 +1,6 @@
+from blue_vault_wallet import BlueVaultWallet
+from generic_ur2_wallet import GenericUR2Wallet
+from seed_storage import SeedStorage
+from sparrow_multisig_wallet import SparrowMultiSigWallet
+from specter_desktop_multisig_wallet import SpecterDesktopMultisigWallet
+from specter_desktop_single_sig_wallet import SpecterDesktopSingleSigWallet
