+++ src/seedsigner/models/seed_storage.py
@@ -1,6 +1,5 @@
 # External Dependencies
-from embit.bip39 import mnemonic_to_bytes
-from embit.bip39 import mnemonic_from_bytes
+from embit.bip39 import mnemonic_to_bytes, mnemonic_from_bytes
 from embit import bip39
 
 class SeedStorage:
