+++ src/seedsigner/models/specter_desktop_multisig_wallet.py
@@ -1,15 +1,10 @@
-from wallet import Wallet
-from qr import QR
+from . import Wallet
+from seedsigner.helpers import QR
 
 # External Dependencies
-from embit.bip39 import mnemonic_to_bytes
-from embit.bip39 import mnemonic_from_bytes
-from embit import bip39
-from embit import script
-from embit import bip32
-from embit import psbt
+from embit import bip32, bip39, ec, psbt, script
+from embit.bip39 import mnemonic_to_bytes, mnemonic_from_bytes
 from embit.networks import NETWORKS
-from embit import ec
 from io import BytesIO
 from binascii import unhexlify, hexlify, a2b_base64, b2a_base64
 import re
