+++ src/seedsigner/views/__init__.py
@@ -1,0 +1,6 @@
+from io_test_view import IOTestView
+from menu_view import MenuView
+from seed_tools_view import SeedToolsView
+from settings_tools_view import SettingsToolsView
+from signing_tools_view import SigningToolsView
+from view import View
