+++ src/seedsigner/views/seed_tools_view.py
@@ -1,12 +1,12 @@
 # External Dependencies
-from embit.bip39 import mnemonic_to_bytes
-from embit.bip39 import mnemonic_from_bytes
 from embit import bip39
+from embit.bip39 import mnemonic_to_bytes, mnemonic_from_bytes
 import time
 
 # Internal file class dependencies
-from view import View
-from buttons import B
+from . import View
+from seedsigner.helpers import B
+
 
 class SeedToolsView(View):
